Refactor/datepick 227

+++ src/main/java/app/hdj/datepick/domain/tag/repository/TagRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Byte> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
 
 }
